Add is_leaf and filter function to jsk_tools'library
